# docs/static/swagger.json
{
...
  "info": {
    "description": "Documenta\u00e7\u00e3o autom\u00e1tica das rotas do sistema com arquitetura DDD e SQL Puro.",
    "title": "Documenta\u00e7\u00e3o da API - Projeto g03",
    "version": "1.0.0"
  },
  "paths": {
    "/api/auth/login": {
      "post": {
        "description": "N\u00e3o exige token. Recebe email e senha, e devolve o token de acesso.<br/>",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "email": {
                  "example": "maria@empresa.com",
                  "type": "string"
                },
                "password": {
                  "example": "senha_segura123",
                  "type": "string"
                }
              },
              "required": [
                "email",
                "password"
              ],
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login bem-sucedido, token retornado",
            "schema": {
              "properties": {
                "message": {
                  "example": "Login realizado com sucesso!",
                  "type": "string"
                },
                "role": {
                  "example": "gestor_remoto",
                  "type": "string"
                },
                "token": {
                  "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                  "type": "string"
                }
              },
              "type": "object"
            }
          },
          "400": {
            "description": "Campos obrigat\u00f3rios ausentes ou formato inv\u00e1lido"
          },
          "401": {
            "description": "E-mail ou senha incorretos"
          }
        },
        "summary": "Autentica um usu\u00e1rio e retorna o token JWT",
        "tags": [
          "Autentica\u00e7\u00e3o"
        ]
      }
    },
    "/api/auth/register": {
      "post": {
        "description": "Rota protegida: apenas gestores remotos autenticados podem criar novos usu\u00e1rios.<br/>",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "email": {
                  "example": "maria@empresa.com",
                  "type": "string"
                },
                "name": {
                  "example": "Maria Souza",
                  "type": "string"
                },
                "password": {
                  "example": "senha_segura123",
                  "type": "string"
                },
                "role": {
                  "description": "Opcional. Se omitido, o usu\u00e1rio ser\u00e1 criado como gestor_local.",
                  "example": "gestor_local",
                  "type": "string"
                }
              },
              "required": [
                "name",
                "email",
                "password"
              ],
              "type": "object"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Usu\u00e1rio criado com sucesso",
            "schema": {
              "properties": {
                "id": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          },
          "400": {
            "description": "Dados inv\u00e1lidos ou e-mail j\u00e1 cadastrado"
          },
          "401": {
            "description": "Token JWT ausente ou inv\u00e1lido"
          },
          "403": {
            "description": "Apenas gestores remotos podem cadastrar usu\u00e1rios"
          },
          "500": {
            "description": "Erro interno ao criar o usu\u00e1rio"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Cadastra um novo usu\u00e1rio no sistema",
        "tags": [
          "Autentica\u00e7\u00e3o"
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "description": "Insira o token JWT no formato: Bearer {seu_token}",
      "in": "header",
      "name": "Authorization",
      "type": "apiKey"
    }
  },
  "swagger": "2.0"
}